Setting getter

Typcally this should not be used directly. It is relied upon by the get_* functions in this package. It is exported mostly to provide transparency around the available parameters.

get_setting(varstr, value, allow_null = FALSE, param, scope, setter, ...)

Arguments

varstr
Character. The name of the cached variable.

value
Expected value (any type). If provided it overrides the cached value.

allow_null
Logical. If FALSE a helpful error will result if the return value is NULL. Otherwise the NULL will be returned silently.

param
Character. Configuration parameter name. See get_cluster_param.

scope
Character. Configuration scope. See get_cluster_param

setter
Function. The set_* function used to keep track of the parameter value. This is needed so that in cases where a parameter is asked for that has not yet been set, the configuration lookup need only happen once and then the value is cached.

...
Passed to get_cluster_param.