Build an active URL for WebHDFS

Full WebHDFS URL is built dynamically from an active namenode for the current cluster.

get_webhdfs_url()

Value

Character containing full WebHDFS URL from an active cluster namenode

Details

It is possible to define an array including primary and backup namenodes for use by the clusterconf package. This function will check the provided namenodes until an active one is found. The full WebHDFS URL is then built from the active namenode, along with the cluster settings for WebHDFS port number and URL suffix. This should provide the user with consistent access to WebHDFS services, regardless of which namenode is currently active.

Examples

# NOT RUN {
set_cluster("my prefered cluster")
get_webhdfs_url()
# }