Generic functions to perform HTTP GET or PUT requests of a WebHDFS operation

The full URL is built from the base WebHDFS URL via the get_webhdfs_url function, and the supplied path and operation parameters.

hdfs_get(path, operation, user = get_user(),
  return_type = get_return_type(), handler = NULL)

hdfs_put(path, operation, user = get_user())

Arguments

path
Character containing file system path

operation
Character containing WebHDFS operation name. This can include the operation itself, along with additional parameters as necessary. If parameters are included, they should be separated from the operation and each other with '&', such as 'OPERATION&param1=value1&param2=value2'.

user
Character username to use in WebHDFS operation. If not provided, webhdfs.user will be used and if that has not been set, a call to guess_user will be made.

return_type
character string. See set_return_type for details and options.

handler
Function to use when processing the result of the WebHFDS operation. If not provided, defaults to jsonlite::fromJSON, which handles output from the vast majority of GET operations.

Value

data.frame (or other requested type) containing output from the WebHDFS operation

Details

If the WebHDFS command returns an exception, this function issues a warning containing the exception message and returns an empty data.frame.

See also

http://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-hdfs/WebHDFS.html for documentation of WebHDFS REST API commands that can be passed to the operation parameter

Examples

# NOT RUN {
# basic usage with operation
hdfs_get("/data/", "LISTSTATUS")

# usage with operation appended to parameter
hdfs_put("/user/uid/newdir", "MKDIRS&permission=754")
# }