List a directory

Shortcut for WebHDFS LISTSTATUS operation

hdfs_ls(path, recursive = FALSE, concise = FALSE, user = get_user(),
  return_type = get_return_type())

clean_liststatus_columns(dat)

Arguments

path
Character containing file system path

recursive
Logical indicator of whether to resursively list individual files within sub-directories. Default FALSE.

concise
Logical indicator of whether to return only a select subset of columns. Default FALSE.

user
Character username to use in WebHDFS operation. If not provided, webhdfs.user will be used and if that has not been set, a call to guess_user will be made.

return_type
character string. See set_return_type for details and options.

dat
Data frame containing the result of a LISTSTATUS command

Value

data.frame (or other requested type) containing output from the WebHDFS operation

Details

Wildcards (*) in the provided path are handled by getting all contents at that level of the directory structure, and appending any remaining portions of the path below that level.

When concise = TRUE, function clean_liststatus_columns is called to limit the set of columns from the WebHDFS LISTSTATUS operation to the following fields in this order: pathSuffix, childrenNum, length, group, modificationTime, owner, permission, type. Note that this places the pathSuffix field as the first column, which differs from the default ordering. When concise = FALSE, the results include the full set of columns returned by LISTSTATUS, in the original order.

Examples

# NOT RUN {
# basic usage to list directory contents
hdfs_ls("/data/")

# recursively list files within sub-directories
hdfs_ls("/data/blah/2016", recursive = TRUE)

# use wildcard to get the first day of every month in 2016
hdfs_ls("/data/blah/2016/*/01")

# use wildcards to get all days in July in every year
hdfs_ls("/data/blah/*/07/*")
# }