Open and read a file

Shortcut to the WebHDFS OPEN command.

hdfs_open(path, handler, offset_bytes = NULL, length_bytes = NULL,
  buffer_size = NULL, user = get_user(), return_type = get_return_type())

Arguments

path
Character containing file system path

handler
Function name to be used to load data from WebHDFS. It is incumbent on the user for this function to be available in the current session, and for the function to handle the type of data referenced by path. This is passed on to hdfs_get.

offset_bytes
The starting position to be processed, in bytes. Default (NULL) means start from the beginning of the file.

length_bytes
The number of bytes to be processed. Default (NULL) means the entire file will be read.

buffer_size
The size of the buffer to be used in transferring data.

user
Character username to use in WebHDFS operation. If not provided, webhdfs.user will be used and if that has not been set, a call to guess_user will be made.

return_type
character string. See set_return_type for details and options.

Value

data.frame (or other requested type) containing content from the requested file path

See also

http://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-hdfs/WebHDFS.html#Open_and_Read_a_File for documentation of WebHDFS OPEN command and its parameters.

Examples

# NOT RUN {
library(data.table)
dat <- hdfs_open("/data/dummy.csv", handler = "fread")
# }