Specify a Hadoop cluster

The easy way to set started - skipping the call to many different set_ functions.

set_cluster(name, config_yaml)

get_cluster()

Arguments

name
Character. Specify the name of the cluster to "point" to.

config_yaml
Character. Path to a YAML file specifying cluster settings.

Details

This function will take either a cluster name or a path to a configuration file. If a name is provided then a search will be made for a package containing cluster configurations (named clusterconf.<your cluster name>). The search starts in you installed packages (.libPaths()) and then proceeds to check your configured repositories (getOption("repos")).

If you provide a configuration file path instead then that is used in lieu of searching for a package and the necessary settings are parsed out of it.