Unlist an object to the innermost data.frame

Some calls to the WebHDFS REST API return a single JSON object, and some return an array of JSON objects. This is a general function that will return a data.frame based on the output of either type from fromJSON

unlist_carefully(x)

Arguments

x
Object returned from call to fromJSON, typically a list

Value

data.frame containing the innermost data.frame-like object within x

Details

This is similar to simply calling as.data.frame on a list, but without the nested column names that result. For example, where as.data.frame could generate a column name of FileStatuses.FileStatus.accessTime, this function will result in a column name of simply accessTime.

Examples

# sample data.frame nested within multiple lists
test <- list(outer = list(inner = data.frame(x = 1, y = 1:10)))
test

#> $outer
#> $outer$inner
#>    x  y
#> 1  1  1
#> 2  1  2
#> 3  1  3
#> 4  1  4
#> 5  1  5
#> 6  1  6
#> 7  1  7
#> 8  1  8
#> 9  1  9
#> 10 1 10
#> 
#> 

# this function unpacks to a simple data.frame
unlist_carefully(test)

#>    x  y
#> 1  1  1
#> 2  1  2
#> 3  1  3
#> 4  1  4
#> 5  1  5
#> 6  1  6
#> 7  1  7
#> 8  1  8
#> 9  1  9
#> 10 1 10

# compare that to the behavior of as.data.frame
as.data.frame(test)

#>    outer.inner.x outer.inner.y
#> 1              1             1
#> 2              1             2
#> 3              1             3
#> 4              1             4
#> 5              1             5
#> 6              1             6
#> 7              1             7
#> 8              1             8
#> 9              1             9
#> 10             1            10